+++ libraries/_helper.rb
@@ -74,7 +74,8 @@ If this problem persists, check your Jenkins log files.
         h[:protocol] = protocol unless protocol.nil?
         h[:cli_user] = cli_user unless cli_user.nil?
         h[:cli_username] = cli_username unless cli_username.nil?
-        h[:cli_password] = cli_password unless cli_password.nil?
+        h[:cli_apitoken] = cli_apitoken unless cli_apitoken.nil?
+        h[:cli_credential_file] = cli_credential_file unless cli_credential_file.nil?
       end
 
       Jenkins::Executor.new(options)
@@ -393,12 +394,21 @@ If this problem persists, check your Jenkins log files.
     end
 
     #
-    # password to pass to cli
+    # cli_apitoken to pass to cli
     #
     # @return [String]
     #
-    def cli_password
-      node['jenkins']['executor']['cli_password']
+    def cli_apitoken
+      node['jenkins']['executor']['cli_apitoken']
+    end
+
+    #
+    # cli_credential_file to pass to cli
+    #
+    # @return [String]
+    #
+    def cli_credential_file
+      node['jenkins']['executor']['cli_credential_file']
     end
 
     #